% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_unit.R
\name{convert_unit}
\alias{convert_unit}
\title{Data Size Unit Conversion}
\usage{
convert_unit(
  size,
  in_unit = c("B", "KB", "MB", "GB", "TB", "PB"),
  out_unit = c("AUTO", "B", "KB", "MB", "GB", "TB", "PB")
)
}
\arguments{
\item{size}{Data size.}

\item{in_unit}{Input data unit.}

\item{out_unit}{Output data unit. This can be set to "AUTO" for automatic optimization, which is useful for plotting.}
}
\value{
Converted data size to output unit.
}
\description{
Convert storage size data to different units.
}
\examples{
convert_unit(10000, "KB", "MB")
convert_unit(15007, "KB", "AUTO")
}
