% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_multielement_groups.R
\name{detect_multielement_groups}
\alias{detect_multielement_groups}
\title{Detect Multi-Element Groups}
\usage{
detect_multielement_groups(tags)
}
\arguments{
\item{elements}{String vector with tags.}
}
\value{
Vector with numeric group indexes.
}
\description{
Detect groups with multiple elements and give them a common index number.
}
\examples{
detect_multielement_groups(c("asd.txt","asd.Rdata", "qwe.txt","rrrrrrrr.txt","asd.txt", "asd.txt"))
}
