% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_multielement_subgroups.R
\name{detect_multielement_subgroups}
\alias{detect_multielement_subgroups}
\title{Detect Multi-Element Subgroups}
\usage{
detect_multielement_subgroups(subgroup_property, groups_midx)
}
\arguments{
\item{subgroup_property}{Vector (numeric or character) with a secondary group property}

\item{groups_midx}{Numeric vector with group indexes indicating instaces to be compared against each other.}
}
\value{
Numeric vector with indexes for subgroups with the same subroup property.
}
\description{
Detect subgroups inside groups and give them a subgroup index case they have more than 1 element.
}
\examples{
detect_multielement_subgroups(c(45, 5, 33, 657, 45, 67), c(1 ,0 ,0 ,0 ,1 ,1))
}
