% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_related_files.R
\name{detect_related_files}
\alias{detect_related_files}
\title{Detect Related Files}
\usage{
detect_related_files(all_fullfn)
}
\arguments{
\item{all_fullfn}{Vector with full file names (file name + extension).}
}
\value{
Matrix with group indexes for related files.
}
\description{
Detect related files (criteria: same name but extension can differ).
}
\examples{
fullfn = c("./myfolder/CopywriteR.log", "CopywriteR.log", "./dir/resume.txt" , "./dif_dir/difsubdir/resume.txt", "single.txt", "CopywriteR.log")
detect_related_files(fullfn)
}
