% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_same.R
\name{detect_same}
\alias{detect_same}
\title{Detect Files With Same Full Name (Name+Extension)}
\usage{
detect_same(all_fullfn)
}
\arguments{
\item{all_fullfn}{Vector with full file names (file name + extension).}
}
\value{
Vector with numeric indexes for groups of files with the same full name.
}
\description{
Detect files with same full name (name+extension) and give them a common index number.
}
\examples{
detect_same(c("asd.txt","asd.Rdata", "qwe.txt","rrrrrrrr.txt","asd.txt", "asd.txt"))
}
