% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_same2.R
\name{detect_same2}
\alias{detect_same2}
\title{Detect Files With Same Full Name (Name+Extension) And Size}
\usage{
detect_same2(size, fullfn_midx)
}
\arguments{
\item{size}{Vector with file sizes.}

\item{fullfn_midx}{Vector with numeric indexes for which size values to comapre against each other.}
}
\value{
Vector with numeric indexes for groups with the same full size value.
}
\description{
Detect files (sharing the same full name) which have the same size and give them a common index number.
}
\examples{
detect_same2(c(45, 5, 33, 657, 45, 67), c(1 ,0 ,0 ,0 ,1 ,1))
}
