% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_same_fullname.R
\name{detect_same_fullname}
\alias{detect_same_fullname}
\title{Detect Files With Same Full Name (Name+Extension)}
\usage{
detect_same_fullname(fullnames)
}
\arguments{
\item{fullnames}{String vector with file full names (name+extension).}
}
\value{
Vector with numeric indexes for groups of files with the same full name.
}
\description{
Detect files with same full name (name+extension) and give them a common index number.
}
\examples{
detect_same_fullname(c("asd.txt","asd.Rdata", "qwe.txt","rrrrrrrr.txt","asd.txt", "asd.txt"))
}
