% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_same_fullname_size.R
\name{detect_same_fullname_size}
\alias{detect_same_fullname_size}
\title{Detect Files With Same Full Name (Name+Extension) And Size}
\usage{
detect_same_fullname_size(sizes, fullname_midx)
}
\arguments{
\item{sizes}{Numeric vector with file sizes.}

\item{fullname_midx}{Numeric vector with group indexes indicating file sizes to be compared against each other.}
}
\value{
Numeric vector with indexes for groups of files with the same full name and size value.
}
\description{
Detect files sharing the same full name and size and give them a group index.
}
\examples{
detect_same_fullname_size(c(45, 5, 33, 657, 45, 67), c(1 ,0 ,0 ,0 ,1 ,1))
}
