% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude.R
\name{exclude}
\alias{exclude}
\title{Exclude Data}
\usage{
exclude(tags, exclude_tags = NULL)
}
\arguments{
\item{tags}{String vector with tags to be filtered.}

\item{exclude_tags}{String vector with tags to exclude from the dataset.}
}
\value{
Bolean flag with FALSE for items to be excluded, and TRUE otherwise.
}
\description{
Detects tags in a set starting with a given exclude tag.
}
\examples{
exclude(c("asdttttttt", "sssasd", "asd", "fff"), c("asd","fffffff"))
}
