% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_counts.R
\name{get_counts}
\alias{get_counts}
\title{Count Total}
\usage{
get_counts(tags, target_tags = NULL)
}
\arguments{
\item{tags}{String vector with tags for which to get counts.}

\item{target_tags}{String vector with tags to keep.}
}
\value{
Table with counts for unique tags.
}
\description{
Calculate total number of instances in a set.
}
\examples{
get_counts(c("txt", "gz", "txt", "gz", "txt", "bam"))
}
