% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_extension.R
\name{get_extension}
\alias{get_extension}
\title{Get File Extension (Single File)}
\usage{
get_extension(dir_fullfn)
}
\arguments{
\item{dir_fullfn}{Directories and full file name (file name + extension) to parse.}
}
\value{
Vector of extensions with parsed extensions.
}
\description{
Get extension for one input file, assuming that the extension follows a standard structure where the extension is the last piece of the file name after a dot.
}
\examples{
get_extension("MyAmazinfFile.pdf"))
}
