% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_extensions.R
\name{get_extensions}
\alias{get_extensions}
\title{Get File Extension}
\usage{
get_extensions(dir_fullfns)
}
\arguments{
\item{dir_fullfns}{Vector with directories and full file names (file name + extension).}
}
\value{
Vector with parsed extensions.
}
\description{
Get extension for files, assuming that the extension follows a standard structure where the extension is the last piece of the file name after a dot.
}
\examples{
get_extensions(c("./CNV/CNAprofiles_HUB.01.B2.001/CopywriteR.log", "./Results/resume.txt"))
get_extensions("MyAmazinfFile.pdf")
}
