% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_formats.R
\name{get_formats}
\alias{get_formats}
\title{Determine File Format}
\usage{
get_formats(extensions)
}
\arguments{
\item{extensions}{Vector with file extensions.}
}
\value{
Vector with all possible file extensions for the known formats. File extensions wich are not recognized (and for which a format cannot be determined) are signed with an "Unknown" tag.
}
\description{
Determine file format from its extension.
}
\examples{
get_format(c("htm", "txt","random_tag"))
get_format("mpg")
}
