% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fullname.R
\name{get_fullname}
\alias{get_fullname}
\title{Get File Full Name (Single File)}
\usage{
get_fullname(dir_fullfn)
}
\arguments{
\item{dir_fullfn}{String tag composed of directoriy followed by file full name (file name + extension).}
}
\value{
Parsed file full name (file name+extension)).
}
\description{
Get file full name (name+extension) from one tag composed of file directory+name+extension.
}
\examples{
get_fullname("MyAmazinfFile.pdf")
}
