% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fullnames.R
\name{get_fullnames}
\alias{get_fullnames}
\title{Get File Full Name}
\usage{
get_fullnames(files)
}
\arguments{
\item{files}{String vector containing for files the path and full name.}
}
\value{
Vector of strings with file full names.
}
\description{
Get file full name (name+extension) from a set of strings by removing all of the path up to and including the last path separator (if any).
}
\examples{
get_fullnames(c("./CNV/CNAprofiles_HUB.01.B2.001/CopywriteR.log", "./Results/resume.txt"))
get_fullnames("MyAmazinfFile.pdf")
}
