% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_name.R
\name{get_name}
\alias{get_name}
\title{Get File Name (Single File)}
\usage{
get_name(fullfn)
}
\arguments{
\item{fullfn}{String to be parsed.}
}
\value{
File name without file extension.
}
\description{
Get file name (file name without extension) from a full file name tag. Assumes that the tag follows the format: <file_name>.<file_extension>.
}
\examples{
get_name("myfile.txt")
}
