% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names.R
\name{get_names}
\alias{get_names}
\title{Get File Name}
\usage{
get_names(files)
}
\arguments{
\item{files}{String vector containing files full name, concatenated or not with the path.}
}
\value{
String vector with file names without extension.
}
\description{
Get file name (name without extension) from a file full name tag for a set of files. Assumes that the tags follows the format: <name>.<extension>.
}
\examples{
get_names(c("/mnt/amazin_project/amazing_file.txt", "bigdata.Rdata"))
}
