% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_progressbar.R
\name{get_progressbar}
\alias{get_progressbar}
\title{Progress Bar}
\usage{
get_progressbar(total)
}
\arguments{
\item{total}{Total number of computer iterations for completion.}
}
\description{
Initialize progress bar.
}
\examples{
get_progressbar(50)
}
