% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_size.R
\name{get_size}
\alias{get_size}
\title{Compute Totals For Sets}
\usage{
get_size(tag, sizes, tags)
}
\arguments{
\item{tag}{Tag for which to compute totals.}

\item{sizes}{Vector with values per tag.}

\item{tags}{Vector with all tags for which a value is provided.}
}
\value{
Total value.
}
\description{
Calculate total from a set of values filtered by corresponding tag.
}
\examples{
get_size("txt", c(10, 50, 10, 33, 77), c("txt", "txt", "sz", "bam", "txt"))
}
