% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sizes.R
\name{get_sizes}
\alias{get_sizes}
\title{Compute Totals For Sets (Fast Version)}
\usage{
get_sizes(sizes, tags)
}
\arguments{
\item{sizes}{Vector with values per tag.}

\item{tags}{Vector with all tags for which a value is provided.}
}
\value{
Vector with the totals for each tag.
}
\description{
Calculate total from a set of values filtered by corresponding tag.
}
\examples{
get_sizes(c(10, 50, 10, 33, 77), c("txt", "txt", "sz", "bam", "txt"))
}
