% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summaries.R
\name{get_summaries}
\alias{get_summaries}
\title{Data Summary}
\usage{
get_summaries(sizes, ids)
}
\arguments{
\item{sizes}{Vector with numbers.}

\item{ids}{Vector with tags.}
}
\value{
Data summaries per unique tag as compute per R function "summary()": Min., 1st Qu.,  Median, Mean, 3rd Qu. and  Max.
}
\description{
Compute data summaries for multiple tags.
}
\examples{
get_summaries(c(10, 10, 33333, 3, 12), c("txt","txt","bam","txt","bam"))
}
