% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary.R
\name{get_summary}
\alias{get_summary}
\title{Data Summary (Single Category)}
\usage{
get_summary(target_id, sizes, ids)
}
\arguments{
\item{target_id}{Tag for which to compute the data summary.}

\item{sizes}{Vector with numbers.}

\item{ids}{Vector with tags.}
}
\value{
Data summary for one tag as compute per R function "summary()": Min., 1st Qu.,  Median, Mean, 3rd Qu. and  Max.
}
\description{
Compute data summary for single tag.
}
\examples{
get_summary("bam", c(10, 10, 33333, 3, 12), c("txt","txt","bam","txt","bam"))
}
