% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top.R
\name{get_top}
\alias{get_top}
\title{Top Ranking}
\usage{
get_top(sizes, ids, max_ids)
}
\arguments{
\item{sizes}{Vector of values.}

\item{ids}{Identifier tags for each input value.}

\item{max_ids}{Number of top elements to return.}
}
\value{
Vector with top values sorted is descending order.
}
\description{
Retrieve categories with top values.
}
\examples{
get_top(c(10, 100, 34, 22, 450), c("txt", "gz", "pdf", "rdata", "bam"), 3)
}
