% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include.R
\name{include}
\alias{include}
\title{Include Data}
\usage{
include(tags, include_tags = NULL)
}
\arguments{
\item{tags}{String vector with tags to be filtered.}

\item{include_tags}{String vector with tags to include in the dataset.}
}
\value{
Flag with TRUE bollean value for items to be included.
}
\description{
Detects tags in a set starting with a given include tag.
}
\examples{
include(c("asdttttttt", "sssasd", "asd"), c("asd","fffffff"))
}
