% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{known_db}
\alias{known_db}
\title{Database of known file extensions}
\format{
A data frame with 1717 rows and 3 columns.
\describe{
    \item{Extension}{List of known file extensions grouped by format.}
    \item{Description}{Short description of the file extensions.}
    \item{Category}{File format category.}
    }
}
\source{
{Adapted from https://en.wikipedia.org/wiki/List_of_file_formats}
}
\usage{
known_db
}
\description{
Contains file extensions grouped by file format, a short description of the format and a format category.
}
\examples{
data(known_db) #Lazy loading.
}
\keyword{datasets}
