% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_barplot.R
\name{plot_barplot}
\alias{plot_barplot}
\title{Barplot}
\usage{
plot_barplot(
  sizes,
  tags,
  max_tags = 20,
  barplot_title,
  x_label = "Instance",
  fill_color = "steelblue"
)
}
\arguments{
\item{sizes}{Numeric vector with category sizes.}

\item{tags}{Vector with tags for which to plot the sizes.}

\item{max_tags}{Maximum tags for which to plot.}

\item{barplot_title}{Main title of the barplot.}

\item{x_label}{Label to be used in the x-axis of the plot.}

\item{fill_color}{Color of the bars.}
}
\description{
Barplot for sizes by category.
}
\examples{
sizes = c(10, 500, 250, 3000)
names(sizes) = "MB"
tags = c("txt", "gz", "pdf", "bam")
plot_barplot(sizes, tags, 3, "Size in storage", "File Extension")
}
