% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boxplot.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{Boxplot}
\usage{
plot_boxplot(
  sizes,
  tags,
  max_tags = 20,
  boxplot_title,
  x_label,
  fill_color = "steelblue"
)
}
\arguments{
\item{sizes}{Numeric vector with sizes.}

\item{tags}{String vector with category tags for which to plot the sizes, one per instance.}

\item{max_tags}{Total number of top categories (with higher size) to include in the boxplot.}

\item{boxplot_title}{Main title of the boxplot.}

\item{x_label}{Label to be used in the x-axis of the plot.}

\item{fill_color}{Color of the boxes.}
}
\description{
Boxplot for sizes for instances grouped by category.
}
\examples{
sizes = c(15, 500, 100, 450, 200, 3000, 300, 600)
names(sizes) = "MB"
tags = c("txt", "txt", "txt", "txt", "bam", "bam", "bam", "gz")
plot_boxplot(sizes, tags, 2, "Size in storage", "File Extension")
}
