% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_donut.R
\name{plot_donut}
\alias{plot_donut}
\title{Donut Chart}
\usage{
plot_donut(
  sizes,
  tags,
  fraction_thresh = 0.03,
  donut_title,
  c_palette = "Set3",
  start_angle = 0
)
}
\arguments{
\item{sizes}{Numeric vector with sizes.}

\item{tags}{Vector with tags for which to plot the sizes.}

\item{fraction_thresh}{Threshold for the smaller fractions of the donut to plot.}

\item{donut_title}{Main title of the donut.}

\item{c_palette}{Color pallete to be used int the donut segments.}

\item{start_angle}{Angle to start the plotting.}
}
\description{
Plot donut for size by category.
}
\examples{
sizes = c(10, 500, 60, 3000)
names(sizes) = "MB"
tags = c("txt", "gz", "pdf", "bam")
plot_donut(sizes, tags, 0.4, "Storage used by file extension")
}
