% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_multicopy_files.R
\name{detect_multicopy_files}
\alias{detect_multicopy_files}
\title{Detect Multi-copy Files}
\usage{
detect_multicopy_files(files, sizes, use_md5sum = FALSE)
}
\arguments{
\item{files}{String vector with files path and full name.}

\item{sizes}{Numeric Vector with file sizes.}

\item{use_md5sum}{Bollean indicating if md5sum codes must be used or not.}
}
\value{
Matrix with 2 columns: one containing group indexes for files with same full name, and another containing group indexes for files that further have the same size. Files not matching the criteria are given the numeric index 0.
}
\description{
Detect files with multiple copies based on file full name (name+extension) and size.
}
\examples{
files = rep("D:/Stuff/GITHUBPortfolio/Storage/data/mitra/known_db.rda", 4)
sizes = c(100, 4, 10, 4)
detect_multicopy_files(fullnames, sizes)
}
