% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_multicopy_files.R
\name{detect_multicopy_files}
\alias{detect_multicopy_files}
\title{Detect Multi-copy Files}
\usage{
detect_multicopy_files(fullnames, sizes)
}
\arguments{
\item{fullnames}{String vector with file full names.}

\item{sizes}{Numeric Vector with file sizes.}
}
\value{
Matrix with 2 columns: one containing group indexes for files with same full name, and another containing group indexes for files that further have the same size. Files not matching the criteria are given the numeric index 0.
}
\description{
Detect files with multiple copies based on file full name (name+extension) and size.
}
\examples{
fullnames = c("CopywriteR.log", "CopywriteR.log", "CopywriteR.log", "resume.txt")
sizes = c(4, 4, 100, 20)
detect_multicopy_files(fullnames, sizes)
}
