% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_related_files.R
\name{detect_related_files}
\alias{detect_related_files}
\title{Detect Related Files}
\usage{
detect_related_files(files)
}
\arguments{
\item{files}{String vector containing for files the path and full name.}
}
\value{
Matrix with group indexes for related files.
}
\description{
Detect related files (criteria: same name but extension can differ).
}
\examples{
files = c("./myfolder/CopywriteR.log", "CopywriteR.log", "./dir/resume.txt" , "./dif_dir/difsubdir/resume.txt", "single.txt", "CopywriteR.log")
detect_related_files(files)
}
