% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_counts.R
\name{get_counts}
\alias{get_counts}
\title{Count Total}
\usage{
get_counts(ids, target_ids = NULL)
}
\arguments{
\item{ids}{Ids for which to get counts.}

\item{target_ids}{Ids to keep.}
}
\value{
Table with counts for unique ids.
}
\description{
Calculate total number of instances in a set.
}
\examples{
get_counts(c("txt", "gz", "txt", "gz", "txt", "bam"))
}
