% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_extension.R
\name{get_extension}
\alias{get_extension}
\title{Get File Extension (Single File)}
\usage{
get_extension(file)
}
\arguments{
\item{file}{Directories and full file name (file name + extension) to parse.}
}
\value{
String with file extension.
}
\description{
Get file extension a string assuming that the file naming follows a structure where the extension is the last piece of the file full name after a dot.
}
\examples{
get_extension("./CNV/CNAprofiles_HUB.01.B2.001/CopywriteR.log")
get_extension("MyAmazinfFile.pdf"))
}
