% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_extensions.R
\name{get_extensions}
\alias{get_extensions}
\title{Get File Extension}
\usage{
get_extensions(files)
}
\arguments{
\item{files}{String vector containing for files the path and full name.}
}
\value{
String vector with file extensions.
}
\description{
Get file extension from a set of strings assuming that the file naming follows a structure where the extension is the last piece of the file full name after a dot.
}
\examples{
get_extensions(c("./CNV/CNAprofiles_HUB.01.B2.001/CopywriteR.log", "./Results/resume.txt"))
get_extensions("MyAmazinfFile.pdf")
}
