% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_format.R
\name{get_format}
\alias{get_format}
\title{Determine File Format (Single File)}
\usage{
get_format(extensions)
}
\arguments{
\item{extensions}{String vector with file extensions.}
}
\value{
String vector with all possible extensions for the detected format, for each file. File extensions wich are not recognized (and for which a format cannot be determined) are assigned the tag "Unknown".
}
\description{
Determine file format from extension.
}
\examples{
get_format(c("htm", "txt","random_tag"))
get_format("mpg")
}
