% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_formats.R
\name{get_formats}
\alias{get_formats}
\title{Determine File Format}
\usage{
get_formats(extensions)
}
\arguments{
\item{extensions}{String vector with file extensions.}
}
\value{
String vector with all possible extensions for the detected format, for each file. File extensions wich are not recognized (and for which a format cannot be determined) are assigned the tag "Unknown".
}
\description{
Determine file format from extension.
}
\examples{
get_formats(c("htm", "txt","random_tag"))
get_formats("mpg")
}
