% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fullname.R
\name{get_fullname}
\alias{get_fullname}
\title{Get File Full Name (Single File)}
\usage{
get_fullname(files)
}
\arguments{
\item{files}{String containing for a file the path and full name.}
}
\value{
String with file full name.
}
\description{
Get file full name (name+extension) from a string by removing all of the path up to and including the last path separator (if any).
}
\examples{
get_fullname("/mnt/folder1/MyAmazinfFile.pdf")
}
