% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fullnames.R
\name{get_fullnames}
\alias{get_fullnames}
\title{Get File Full Name}
\usage{
get_fullnames(dir_fullfns)
}
\arguments{
\item{dir_fullfns}{Vector with tags composed of directory followed by file full name (file name + extension).}
}
\value{
Vector with parsed file full names (file name+extension).
}
\description{
Get file full names (name+extension) from tags composed of directory+name+extension.
}
\examples{
get_fullnames(c("./CNV/CNAprofiles_HUB.01.B2.001/CopywriteR.log", "./Results/resume.txt"))
get_fullnames("MyAmazinfFile.pdf")
}
