% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_name.R
\name{get_name}
\alias{get_name}
\title{Get File Name (Single File)}
\usage{
get_name(file)
}
\arguments{
\item{file}{String containing file full name, concatenated or not with the path.}
}
\value{
File name without extension.
}
\description{
Get file name (name without extension) from a file full name tag. Assumes that the tag follows the format: <name>.<extension>.
}
\examples{
get_name("myfile.txt")
}
