% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names.R
\name{get_names}
\alias{get_names}
\title{Get File Name}
\usage{
get_names(fullfns)
}
\arguments{
\item{fullfns}{Vector with strings to be parsed.}
}
\value{
Vector with filenames without extension.
}
\description{
Get file name (file name without extension) from a full file name tag for a set of files. Assumes that the tag follows the format: <file_name>.<file_extension>.
}
\examples{
get_names(c("amazing_file.txt", "bigdata.Rdata"))
}
