% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top.R
\name{get_top}
\alias{get_top}
\title{Top Ranking}
\usage{
get_top(sizes, tags, max_tags)
}
\arguments{
\item{sizes}{Numeric vector with sizes.}

\item{tags}{String with tags for each size value.}

\item{max_tags}{Number of top elements to return.}
}
\value{
Numeric vector with top values sorted is descending order named by tag.
}
\description{
Retrieve categories with top values.
}
\examples{
get_top(c(10, 100, 34, 22, 450), c("txt", "gz", "pdf", "rdata", "bam"), 3)
}
